requirements.txt
fastapi==0.95.2
uvicorn==0.24.0
sqlalchemy>=1.4,<2.0
psycopg2-binary==2.9.11
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic>=1.8.0,<2.0
python-dotenv==1.0.0